+++ plugins/nexus-repository-manager/dev/index.tsx
@@ -1,39 +1,37 @@
 import React from 'react';
 
-import { Entity } from '@backstage/catalog-model';
 import { createDevApp } from '@backstage/dev-utils';
 import { EntityProvider } from '@backstage/plugin-catalog-react';
+import { TestApiProvider } from '@backstage/test-utils';
 
+import {
+  entityMock,
+  NexusRepositoryManagerApiClientMock,
+} from '../src/__fixtures__/mocks';
+import { NexusRepositoryManagerApiRef } from '../src/api';
 import {
   NexusRepositoryManagerPage,
   nexusRepositoryManagerPlugin,
 } from '../src/plugin';
 
-const mockEntity: Entity = {
-  apiVersion: 'backstage.io/v1alpha1',
-  kind: 'Component',
-  metadata: {
-    name: 'backstage',
-    description: 'backstage.io',
-    annotations: {
-      'nexus-repository-manager/docker.image-name':
-        'janus-idp/backstage-showcase',
-    },
-  },
-  spec: {
-    lifecycle: 'production',
-    type: 'service',
-    owner: 'user:guest',
-  },
-};
-
 createDevApp()
   .registerPlugin(nexusRepositoryManagerPlugin)
   .addPage({
     element: (
-      <EntityProvider entity={mockEntity}>
-        <NexusRepositoryManagerPage />
-      </EntityProvider>
+      <TestApiProvider
+        apis={[
+          [
+            NexusRepositoryManagerApiRef,
+            new NexusRepositoryManagerApiClientMock(
+              require('../src/__fixtures__/components/all.json'),
+            ),
+          ],
+        ]}
+      >
+        <EntityProvider entity={entityMock}>
+          <NexusRepositoryManagerPage />
+        </EntityProvider>
+      </TestApiProvider>
     ),
     title: 'Root Page',
     path: '/nexus-repository-manager',
