+++ plugins/nexus-repository-manager/src/__fixtures__/components/all.json
