+++ plugins/nexus-repository-manager/src/__fixtures__/mocks/index.ts
@@ -1,0 +1,4 @@
+export * from './config-api';
+export * from './entity';
+export * from './nexus-repository-manager-api-client';
+export * from './types';
