+++ plugins/nexus-repository-manager/src/__fixtures__/mocks/types.ts
@@ -1,0 +1,3 @@
+export type MockArgs<T> = Partial<{
+  [K in keyof T]: ReturnType<(typeof jest)['fn']>;
+}>;
