+++ plugins/nexus-repository-manager/src/__fixtures__/repository/docker/v2/janus-idp/backstage-showcase/manifests/sha-33dfe6b.json
