+++ plugins/nexus-repository-manager/src/__fixtures__/service/rest/v1/search/404.json
