+++ plugins/nexus-repository-manager/src/__fixtures__/service/rest/v1/search/docker/imageName/janus-idp/backstage-showcase/index.json
