+++ plugins/nexus-repository-manager/src/__fixtures__/service/rest/v1/search/docker/imageTag/latest/index.json
