+++ plugins/nexus-repository-manager/src/api/index.ts
@@ -1,0 +1,1 @@
+export * from './nexus-repository-manager-api-client';
