--- plugins/nexus-repository-manager/src/classes/NexusRepositoryManagerApiClient/index.ts
@@ -1,1 +1,0 @@
-export * from './NexusRepositoryManagerApiClient';
