--- plugins/nexus-repository-manager/src/classes/index.ts
@@ -1,1 +1,0 @@
-export * from './NexusRepositoryManagerApiClient';
